Apply v2 implementation
