+++ ui/src/main/kotlin/com/processout/sdk/ui/napm/NativeAlternativePaymentBottomSheet.kt
@@ -26,14 +26,12 @@ import com.processout.sdk.ui.napm.NativeAlternativePaymentActivityContract.Compa
 import com.processout.sdk.ui.napm.NativeAlternativePaymentActivityContract.Companion.EXTRA_RESULT
 import com.processout.sdk.ui.napm.NativeAlternativePaymentCompletion.Failure
 import com.processout.sdk.ui.napm.NativeAlternativePaymentCompletion.Success
+import com.processout.sdk.ui.napm.NativeAlternativePaymentEvent.Dismiss
+import com.processout.sdk.ui.napm.NativeAlternativePaymentEvent.PermissionRequestResult
 import com.processout.sdk.ui.napm.NativeAlternativePaymentSideEffect.PermissionRequest
+import com.processout.sdk.ui.napm.NativeAlternativePaymentViewModelState.Loaded
+import com.processout.sdk.ui.napm.NativeAlternativePaymentViewModelState.Stage
 import com.processout.sdk.ui.napm.PONativeAlternativePaymentConfiguration.Flow
-import com.processout.sdk.ui.napm.v2.NativeAlternativePaymentEvent.Dismiss
-import com.processout.sdk.ui.napm.v2.NativeAlternativePaymentEvent.PermissionRequestResult
-import com.processout.sdk.ui.napm.v2.NativeAlternativePaymentScreen
-import com.processout.sdk.ui.napm.v2.NativeAlternativePaymentViewModel
-import com.processout.sdk.ui.napm.v2.NativeAlternativePaymentViewModelState.Loaded
-import com.processout.sdk.ui.napm.v2.NativeAlternativePaymentViewModelState.Stage
 import com.processout.sdk.ui.shared.component.displayCutoutHeight
 import com.processout.sdk.ui.shared.component.screenModeAsState
 import com.processout.sdk.ui.shared.configuration.POBottomSheetConfiguration.Height.Fixed
